refactor(pipeline): extract _models.py from ingest_batch _core.py (#781)
Extract 234 lines of SQL constants, protocol classes, and dataclasses
into ingest_batch/_models.py. _core.py reduced from 1326 to 1123 lines.
Private symbols re-exported in __init__.py for test compatibility.
