+++ docs/topology-status.md
@@ -29,12 +29,12 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 595
+- **Stable** (no move scoped): 596
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 10
 - **TBD** (cell needs explicit assignment): 1
-- **Total declared**: 709
-- **Realized polylogue/**/*.py**: 709 files declared
+- **Total declared**: 710
+- **Realized polylogue/**/*.py**: 710 files declared
 
 ### TBD cells (require explicit routing)
